+++ stable/str.cpp
@@ -432,56 +432,18 @@ template <typename sint, typename uint> sint GenericSignedConversion_(
 
 namespace {
   namespace {
-    namespace {
-      namespace untemplated_ {
-        bso::sBool Conversion(
-          flw::rRFlow &Flow,
-          unsigned long long int &Number,
-          sBase Base,
-          unsigned long long int Limit,
-          unsigned long long int Dummy,
-          qRPN)
-        {
-          return Flow.GetNumber(Number, Base, Limit, qRP);
-        }
-
-        bso::sBool Conversion(
-          flw::rRFlow &Flow,
-          signed long long int &Number,
-          sBase Base,
-          signed long long int PositiveLimit,
-          signed long long int NegativeLimit,
-          qRPN)
-        {
-          return Flow.GetNumber(Number, Base, PositiveLimit, NegativeLimit, qRP);
-        }
-      }
-
-      template <typename type> bso::sBool FlowConversion_(
-        flw::rRFlow &Flow,
-        type &Number,
-        sBase Base,
-        type PositiveLimit,
-        type NegativeLimit,
-        qRPN)
-      {
-        return untemplated_::Conversion(Flow, Number, Base, PositiveLimit, NegativeLimit, qRP);
-      }
-    }
-
-    template <typename type> sdr::sRow StringConversion_(
+    template <typename type, typename limit> sdr::sRow StringConversion_(
       const str::string_ &String,
       type &Number,
-      sdr::row__ Begin,
+      sPos Begin,
       sBase Base,
-      type PositiveLimit,
-      type NegativeLimit,
+      limit Limit,
       qRPN)
     {
       flx::sStringRFlow Flow;
       bso::sBool IsError = false;
 
-      if ( *Begin >= String.Amount() ) {
+      if ( Begin.Value() >= String.Amount() ) {
         if ( qRPT )
           qRFwk();
         else
@@ -489,25 +451,24 @@ namespace {
       }
 
       Flow.Init(String);
-      Flow.Skip(*Begin, &IsError);
+      Flow.Skip(Begin.Value(), &IsError);
 
       if ( IsError )
         qRUnx();
 
-      if ( FlowConversion_(Flow, Number, Base, PositiveLimit, NegativeLimit, qRP) )
+      if ( Flow.GetNumber(Number, Base, Limit, qRP) )
         return qNIL;
       else
         return Flow.AmountRed();
     }
   }
 
-  template <typename type> type StringConversion_(
+  template <typename type, typename limit> type StringConversion_(
     const str::string_ &String,
-    sdr::row__ Begin,
+    sPos Begin,
     sdr::row__ *ErrP,
     sBase Base,
-    type PositiveLimit,
-    type NegativeLimit )
+    limit Limit)
   {
     type Result = 0;
     sdr::sRow ErrorPointer = qNIL;
@@ -515,7 +476,7 @@ namespace {
     if ( ( ErrP != NULL ) && ( *ErrP != qNIL ) )
       return 0;
 
-    if ( ( ErrorPointer = StringConversion_(String, Result, Begin, Base, PositiveLimit, NegativeLimit, qRPU ) ) != qNIL ) {
+    if ( ( ErrorPointer = StringConversion_(String, Result, Begin, Base, Limit, qRPU ) ) != qNIL ) {
       if ( ErrP != NULL )
         *ErrP = ErrorPointer;
       else
@@ -528,24 +489,23 @@ namespace {
 
 unsigned long long int str::UConversion(
 	const str::string_ &String,
-	sdr::row__ Begin,
+	sPos Begin,
 	sdr::row__ *ErrP,
 	sBase Base,
-	unsigned long long int Limit)
+	sULimit<unsigned long long int> Limit)
 {
-  return StringConversion_(String, Begin, ErrP, Base, Limit, 0ULL);
+  return StringConversion_<unsigned long long int, sULimit<unsigned long long int>>(String, Begin, ErrP, Base, Limit);
 }
 
 
 signed long long int str::SConversion(
 	const str::string_ &String,
-	sdr::row__ Begin,
+	sPos Begin,
 	sdr::row__ *ErrP,
 	sBase Base,
-	signed long long int PositiveLimit,
-	signed long long int NegativeLimit)
+	sSLimits<signed long long int> Limits)
 {
-  return StringConversion_(String, Begin, ErrP, Base, PositiveLimit, NegativeLimit);
+  return StringConversion_<signed long long int, sSLimits<signed long long int>>(String, Begin, ErrP, Base, Limits);
 }
 /*
 uint__ str::_UIntConversion(