+++ stable/tol.h
@@ -95,7 +95,7 @@ namespace str {
 
 # define qXENUM( name, prefix )	typedef tol::extended_enum__<e##name##_, prefix##_False, prefix##_Error, prefix##_Undefined> e##name
 
-# define TOL_ROW_( name ) E_TMIMIC__( sdr::bRow, name )
+# define TOL_ROW_( name ) E_TMIMIC__( sdr::tRow, name )
 
 # define qROW( name ) TOL_ROW_( s##name )
 // Nota: although a static object, defining it as resource containing object